Fixed floating point precision

+++ src/Transactions/AddCreditTransaction.cpp
@@ -1,4 +1,4 @@
 #include "AddCreditTransaction.hpp"
 
-AddCreditTransaction::AddCreditTransaction(const std::string &userName, UserType userType, float credits)
+AddCreditTransaction::AddCreditTransaction(const std::string &userName, UserType userType, double credits)
     : BasicTransaction(kTransactionType_AddCredit, userName, userType, credits) {}
