+++ src/Transactions/DeleteTransaction.cpp
@@ -1,4 +1,4 @@
 #include "DeleteTransaction.hpp"
 
-DeleteTransaction::DeleteTransaction(const std::string &userName, UserType userType, float credits)
+DeleteTransaction::DeleteTransaction(const std::string &userName, UserType userType, double credits)
     : BasicTransaction(kTransactionType_Delete, userName, userType, credits) {}
