+++ src/Transactions/LogoutTransaction.cpp
@@ -1,4 +1,4 @@
 #include "LogoutTransaction.hpp"
 
-LogoutTransaction::LogoutTransaction(const std::string &userName, UserType userType, float credits)
+LogoutTransaction::LogoutTransaction(const std::string &userName, UserType userType, double credits)
     : BasicTransaction(kTransactionType_Logout, userName, userType, credits) {}
